Should work this time
I deleted the gauge class and expanded the meter class. Specifically, the meter class now has `relativeValueOfMeterFromTo` and `relativeValueOfMeterWithPoints` methods that calculate the relative value of a meter given a start/end date, or given the number of points to retrieve from the db. The latter method requires that you specify a resolution for the data as it can not be chosen automatically based on the start date. See my gauges repo for examples of both these methods in use.